+++ sdk/monitor/mgmt/src/main/java/com/azure/management/monitor/WorkspaceInfo.java
@@ -35,7 +35,7 @@ public class WorkspaceInfo {
      *
      * @return the id value.
      */
-    public String getId() {
+    public String id() {
         return this.id;
     }