+++ sdk/network/mgmt/src/main/java/com/azure/management/network/implementation/ExpressRouteCircuitPeeringImpl.java
@@ -84,7 +84,7 @@ class ExpressRouteCircuitPeeringImpl
 
     @Override
     public boolean isInCreateMode() {
-        return this.inner().getId() == null;
+        return this.inner().id() == null;
     }
 
     @Override
@@ -104,7 +104,7 @@ class ExpressRouteCircuitPeeringImpl
 
     @Override
     public String id() {
-        return inner().getId();
+        return inner().id();
     }
 
     @Override