+++ sdk/network/mgmt/src/main/java/com/azure/management/network/models/ExpressRouteGatewayInner.java
@@ -126,7 +126,7 @@ public class ExpressRouteGatewayInner extends Resource {
      *
      * @return the id value.
      */
-    public String getId() {
+    public String id() {
         return this.id;
     }