+++ sdk/resources/mgmt/src/main/java/com/azure/management/resources/models/LocationInner.java
@@ -53,7 +53,7 @@ public final class LocationInner {
      *
      * @return the id value.
      */
-    public String getId() {
+    public String id() {
         return this.id;
     }