+++ sdk/sql/mgmt/src/main/java/com/azure/management/sql/implementation/ReplicationLinkImpl.java
@@ -30,7 +30,7 @@ class ReplicationLinkImpl extends RefreshableWrapperImpl<ReplicationLinkInner, R
         this.resourceGroupName = resourceGroupName;
         this.sqlServerName = sqlServerName;
         this.sqlServerManager = sqlServerManager;
-        this.resourceId = ResourceId.fromString(this.inner().getId());
+        this.resourceId = ResourceId.fromString(this.inner().id());
     }
 
     @Override
@@ -154,12 +154,12 @@ class ReplicationLinkImpl extends RefreshableWrapperImpl<ReplicationLinkInner, R
 
     @Override
     public String name() {
-        return this.inner().getName();
+        return this.inner().name();
     }
 
     @Override
     public String id() {
-        return this.inner().getId();
+        return this.inner().id();
     }
 
     @Override