+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 1.0.0-beta.5 (Unreleased)
-
+- Rename `apiKey()` to `credential()` on TextAnalyticsClientBuilder.
 
 ## 1.0.0-beta.4 (2020-04-07)
 - Throws an illegal argument exception when the given list of documents is an empty list.